app/requirements.txt
fastapi
uvicorn[standard]

# Authentication
PyJWT
bcrypt
pydantic[email]
httpx
python-dotenv
sqlalchemy
psycopg2-binary
psycopg[binary,pool]
langgraph-checkpoint-postgres
alembic
asyncpg
pytest
pytest-asyncio

# Google AI Studio
langchain-google-genai
langgraph
langchain


# Etl com excel
pandas
openpyxl

# Monitoring
prometheus-client
boto3

# LangSmith (AI Observability)
langsmith


prometheus_client